requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
boto3==1.21.1
botocore==1.24.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
fastapi==0.72.0
filelock==3.6.0
filetype==1.0.9
Flask==2.0.3
Flask-RESTful==0.3.9
flatbuffers==2.0
frozenlist==1.3.0
gdown==4.2.0
h11==0.13.0
idna==3.3
imageio==2.16.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
llvmlite==0.38.0
MarkupSafe==2.0.1
multidict==6.0.2
networkx==2.6.3
numba==0.55.1
numpy==1.21.5
onnxruntime==1.10.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.0
protobuf==3.19.4
pydantic==1.9.0
PyMatting==1.1.5
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
s3transfer==0.5.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
threadpoolctl==3.1.0
tifffile==2022.2.9
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.0
Werkzeug==2.0.3
wincertstore==0.2
yarl==1.7.2
kafka-python==2.0.2